Remove incorrect conversion from save file logic
The projection enumerable was being converted to a double before saving
to a save file. The save file still seems to store it as a plain
integer, so fixing this shouldn't result in compatibility issues.

+++ src/haloray-core/gui/stateSaver.cpp
@@ -68,7 +68,7 @@ void StateSaver::SaveState(QString filename, SimulationEngine *engine, CrystalPo
 
     settings.beginGroup("Camera");
     auto camera = engine->getCamera();
-    settings.setValue("Projection", (double)camera.projection);
+    settings.setValue("Projection", camera.projection);
     settings.setValue("Pitch", (double)camera.pitch);
     settings.setValue("Yaw", (double)camera.yaw);
     settings.setValue("FieldOfView", (double)camera.fov);
